Support running Fody with Code Contracts
If the Code Contracts extension or NuGet package is used in a project, it does an additional compilation stage and produces a secondary assembly for analysis. Fody needs to run against this extra assembly otherwise the Code Contracts analysis stage will flag warnings for things that are potentially resolved by a Fody addin in the primary output assembly.