added support for different weaver source - #416
This is a per-weaver configuration.
Before looking for in-solution or "Weavers.dll", Fody will look for an attribute named "Source" in the xml weaver element.

Usage: `<ExampleWeaver Source="ExampleSource"/>`
For this very ExampleWeaver, Fody will look for the first reference of ExampleSource.dll
if there are multiple assemblies named ExampleSource, Fody will only search for the weaver in the first it finds. This means this feature is also affected by #415.