Support delay signing through setting public key on ModuleDefinition
* Add support for delay-signing with a public only keyfile

Problem:
Delay-signing allows for keyfiles to contain only the public key since the private key is not needed. Unfortunately, public key bytes do not create a valid StrongNameKeyPair.  StrongNameKeyPair requires the private key to regenerate the public key. Mono.Cecil uses the StrongNameKeyPair to set the public key on the assembly and to sign the assembly.  If an invalid StrongNameKeyPair is given to Mono.Cecil, an exception is thrown and Fody fails.

Fix:
Avoid creating the StrongNameKeyPair and instead use the new PublicKeyBytes on the WriterParameters from Mono.Cecil. If provided, Mono.Cecil will directly set these bytes on the AssemblyName and avoid attempting to sign the assembly.

It's a bit difficult to tell what type of keyfile is present. The simplest way is to create the StrongNameKeyPair and attempt to generate the public key. If this throws an ArgumentException, we know that the private key was not present within the keyfile.  In this case, null out the StrongNameKeyPair and pass along the key file bytes as the public key bytes.