Keep the original file extension in WeaverTestHelper.ExecuteTestRun (#961)
* Update test projects from .NET Core 2.2 to .NET Core 3.1

Also change the condition for referencing Mono.Cecil from $(TargetFramework) to $(TargetFrameworkIdentifier). This makes updating the versions in <TargetFrameworks> easier because the conditions don't need to be updated.

* Keep the original file extension in WeaverTestHelper.ExecuteTestRun

Before this commit, when supplying a custom assembly name, the target file would be renamed with `.dll` when weaving an `.exe` file. This was confusing.